Include marker extension in SGN detection
