+++ scripts/sgn_detection/sgn_detection.py
@@ -14,6 +14,10 @@ def main():
     parser.add_argument("-k", "--input_key", default=None,
                         help="The key / internal path to image data.")
 
+    parser.add_argument("-d", "--extension_distance", type=float, default=12, help="Extension distance.")
+    parser.add_argument("-r", "--resolution", type=float, nargs="+", default=[3.0, 1.887779, 1.887779],
+                        help="Resolution of input in micrometer.")
+
     parser.add_argument("--s3", action="store_true", help="Use S3 bucket.")
     parser.add_argument("--s3_credentials", type=str, default=None,
                         help="Input file containing S3 credentials. "
@@ -28,6 +32,11 @@ def main():
     block_shape = (12, 128, 128)
     halo = (10, 64, 64)
 
+    if len(args.resolution) == 1:
+        resolution = tuple(args.resolution, args.resolution, args.resolution)
+    else:
+        resolution = tuple(args.resolution)
+
     if args.s3:
         input_path, fs = s3_utils.get_s3_path(args.input, bucket_name=args.s3_bucket_name,
                                               service_endpoint=args.s3_service_endpoint,
@@ -37,7 +46,8 @@ def main():
         input_path = args.input
 
     sgn_detection(input_path=input_path, input_key=args.input_key, output_folder=args.output_folder,
-                  model_path=args.model, block_shape=block_shape, halo=halo)
+                  model_path=args.model, block_shape=block_shape, halo=halo,
+                  extension_distance=args.extension_distance, sampling=resolution)
 
 
 if __name__ == "__main__":
